Allow loading of trades page <= min_value when we have unshipped trades.
Note that this may signifigantly increase load time as it now has to
scroll to the bottom of the entire trades page looking for people
we've already got trades outstanding.   If your potential list of
trades is long, that is.

However, I did add some more logic to not bother with traders who have
low points and are NOT in our unshipped list, so the trade dictionary
shouldn't grow too fast/large.